Add the start of the monthly `BSEddies Awards`! Featuring stats and prizes.
Add functionality to trigger a monthly 'stats' message for the server. This triggers on the first day of a month and provides stats on the previous month. There are also some stats that can be 'won' - these are the BSEddies Awards. Users can win these and will gain a small eddie prize.

- Add awards task
- Add stats gathering and caching class
- Add supporting enums and constants for needed vars