#128 - Reduce revolution view complexity
Refactor revolution view to remove a lot of the duplicated code.

All the logic is now abstracted to `_revolution_button_logic` and handles the logic for all buttons. This reduces the amount of duplicated code in the class.
I have also added lots more comments in the hope of making it clearer to understand.
This paves the way for #121; adding a new "neutral" button for supporters/revolutionaries wanting to reset their stance.