#128, #121 - Revolution view refactor (#391)
* #128 - Reduce revolution view complexity

Refactor revolution view to remove a lot of the duplicated code.
All the logic is now abstracted to `_revolution_button_logic` and handles the logic for all buttons. This reduces the amount of duplicated code in the class.
I have also added lots more comments in the hope of making it clearer to understand.

* #121 - Add impartial button
Allow users to be 'impartial'. This means users can make a conscious decision to not care! Either not doing anything at all or resetting their existing choice without actively choosing the other faction.