#393 - Message strings (#394)
Add a central director for some message strings. This is where each file can denote a particular set of messages to be chosen randomly by the bot. Allows simpler editing of some files without having to go looking for their locations.

Will endeavour to add more to this going forward.