Bet updates (#241, #332) (#405)
This update is threefold:
- Allow the creator of the bet to add another option post-creation
- Allow users to change their bets to some extent
- Add some extra top-level fields for bets that will make future easier (#403)

Users will only be able to change their bets 5 minutes after they placed theirs, or five minutes after a new option was added. This is mainly targeting the use case where people place on one outcome accidentally or wish to change it after placing it too quickly without thinking. Adding a new bet option allows users to change their existing bets if they want to change it to the new one. Made it five minutes to prevent users changing a bet after the real outcome becomes clear; though this might be changed in the future to be a certain fraction of the total bet time (longer bets should have more leeway for changing their mind).