#398 - Update to use Dataclasses over TypedDicts and add more comprehensive testing (#547)
This is probably the largest PR being submitted at once - it's a shame that these changes will largely go unnoticed by the userbase and should be relatively seamless. The aim is to make it easier to contribute to the repository and make the codebase more resilient. The former is achieved with better type hints and better consistency when using database interactions. The latter is achieved with better testing (both in quantity and quality) and reducing the number of SonarCloud issues.

Massive refactor to use dataclasses over TypedDicts. Changes:
- changed typed dicts to dataclasses (and refactor the file structure for datatypes)
- refactor mongo classes to remove unnecessary intermediate class
- refactor mongo classes to enforce a `make_data_class` method
- refactor mongo queries to return whatever `make_data_class` returns
- add a "minimum_projection" level to allow us to set minimum projections to ensure our dataclasses can always be instantiated
- update everywhere to use new dataclasses
- add an entire suite of tests for all the dataclasses and database classes using mock data

Other changes unrelated to the main objective:
- remove some unnecessary/deprecated methods that weren't being used
- update README.md and CONTRIBUTING.md
- fix a host of sonar issues across the board
- refactor a few of the classes to reduce complexity where possible
- optimise some of the stats stuff to make less API calls
- make it easier to test/debug the stats stuff
- switch all datetime stuff to UTC
- updated a whole bunch of type hints
- added test suites for other areas of the code as well